screen.cpp: fixed RFeed, things now work!
It was a hasty thinko, I didn't pay enough attention and still don't
fully realize why the asymmetric reverse process helps estimate changes
in the internal state.  I need to give it some thought with pencil and
paper.

As for speed, the program now runs (with the default parameters and
no arguments on the command line) 4.6s user instead of 14.0s user.
A 3x speedup!  (Update: this was measured with -O2; with -O3, it's
11.4s vs 2.46s user time, so more like 4x-5x speedup!)

Furthermore, I checked that the output is exactly the same (apart from
one minor issue in ReportStructure: _v1[0] and _v2[0] are uninitialized,
and probably should not be reported.)

Update: with "./screen 9999" the best result is

    // minVal = 269
    s0 -= data[0];    s2 ^= s10;    s0 = Rot64(s0, 21);    s11 += s0;    s11 ^= s1;
    s1 -= data[1];    s3 ^= s11;    s1 = Rot64(s1, 28);    s0 += s1;    s0 ^= s2;
    s2 -= data[2];    s4 ^= s0;    s2 = Rot64(s2, 33);    s1 += s2;    s1 ^= s3;
    s3 -= data[3];    s5 ^= s1;    s3 = Rot64(s3, 55);    s2 += s3;    s2 ^= s4;
    s4 -= data[4];    s6 ^= s2;    s4 = Rot64(s4, 9);    s3 += s4;    s3 ^= s5;
    s5 -= data[5];    s7 ^= s3;    s5 = Rot64(s5, 14);    s4 += s5;    s4 ^= s6;
    s6 -= data[6];    s8 ^= s4;    s6 = Rot64(s6, 24);    s5 += s6;    s5 ^= s7;
    s7 -= data[7];    s9 ^= s5;    s7 = Rot64(s7, 50);    s6 += s7;    s6 ^= s8;
    s8 -= data[8];    s10 ^= s6;    s8 = Rot64(s8, 28);    s7 += s8;    s7 ^= s9;
    s9 -= data[9];    s11 ^= s7;    s9 = Rot64(s9, 46);    s8 += s9;    s8 ^= s10;
    s10 -= data[10];    s0 ^= s8;    s10 = Rot64(s10, 53);    s9 += s10;    s9 ^= s11;
    s11 -= data[11];    s1 ^= s9;    s11 = Rot64(s11, 29);    s10 += s11;    s10 ^= s0;